Cleanup and formatting

+++ src/dark/mining/client/render/MMBlockRenderingHandler.java
@@ -2,7 +2,6 @@ package dark.mining.client.render;
 
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.IBlockAccess;
 
 import org.lwjgl.opengl.GL11;
@@ -12,10 +11,6 @@ import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
 import cpw.mods.fml.client.registry.RenderingRegistry;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
-import dark.core.client.models.ModelSolarPanel;
-import dark.core.common.CoreRecipeLoader;
-import dark.core.common.DarkMain;
-import dark.core.prefab.ModPrefab;
 import dark.mining.MMRecipeLoader;
 
 @SideOnly(Side.CLIENT)
@@ -23,11 +18,10 @@ public class MMBlockRenderingHandler implements ISimpleBlockRenderingHandler
 {
     private static MMBlockRenderingHandler instance = new MMBlockRenderingHandler();
     public static final int BLOCK_RENDER_ID = RenderingRegistry.getNextAvailableRenderId();
-    
-    
+
     public static MMBlockRenderingHandler instance()
     {
-        if(instance == null)
+        if (instance == null)
         {
             instance = new MMBlockRenderingHandler();
         }
@@ -41,14 +35,14 @@ public class MMBlockRenderingHandler implements ISimpleBlockRenderingHandler
         if (MMRecipeLoader.frackingPipe != null && block.blockID == MMRecipeLoader.frackingPipe.blockID)
         {
             FMLClientHandler.instance().getClient().renderEngine.bindTexture(RenderFrackingPipe.TEXTURE);
-            GL11.glTranslatef((float)0, (float) 1F, (float) 0);
+            GL11.glTranslatef(0, 1F, 0);
             GL11.glScalef(1.0F, -1F, -1F);
             RenderFrackingPipe.model.renderAll();
-        }else
-        if (MMRecipeLoader.laserSentry != null && block.blockID == MMRecipeLoader.laserSentry.blockID)
+        }
+        else if (MMRecipeLoader.laserSentry != null && block.blockID == MMRecipeLoader.laserSentry.blockID)
         {
             FMLClientHandler.instance().getClient().renderEngine.bindTexture(RenderMiningLaser.TEXTURE);
-            GL11.glTranslatef((float)0, (float) 1.7F, (float) 0);
+            GL11.glTranslatef(0, 1.7F, 0);
             GL11.glScalef(1.0F, -1F, -1F);
             GL11.glRotatef(180, 0, 1, 0);
             RenderMiningLaser.model.renderAll();
